Move sign in logic to a re-usable hook
